feat: #WB-1985 create a hook resize + wrapper image and video
