+++ frontend/src/components/Tiptap/Tiptap.tsx
@@ -32,7 +32,6 @@ import {
 import Color from "@tiptap/extension-color";
 import FontFamily from "@tiptap/extension-font-family";
 import Highlight from "@tiptap/extension-highlight";
-import Image from "@tiptap/extension-image";
 import Link from "@tiptap/extension-link";
 import Subscript from "@tiptap/extension-subscript";
 import Superscript from "@tiptap/extension-superscript";
@@ -56,13 +55,16 @@ import { AttachReact, TestAttachment } from "./AttachmentReact";
 import ImageEditMenu from "./ImageEditMenu";
 import { LinkerNodeView } from "./LinkerNodeView";
 import LinkerToolbar from "./LinkerToolbar";
-import { ImageResize, ResizableMedia } from "./ResizableMedia";
 import TableToolbar from "./TableToolbar";
+import { ImageResize, WrapperResizeImage } from "./WrapperResizeImage";
 import { useActionOptions } from "~/hooks/useActionOptions";
+// eslint-disable-next-line import/order
 import { useToolbarItems } from "~/hooks/useToolbarItems";
 
 import "katex/dist/katex.min.css";
 import "~/styles/table.scss";
+// eslint-disable-next-line import/order
+import { VideoResize, WrapperResizeVideo } from "./WrapperResizeVideo";
 
 export interface TiptapProps {
   appCode?: string;
@@ -114,9 +116,8 @@ const Tiptap = () => {
       AttachReact(TestAttachment),
       LinkerNodeView,
       Hyperlink,
-      ImageResize(ResizableMedia),
-      Image,
-      Video,
+      ImageResize(WrapperResizeImage),
+      VideoResize(WrapperResizeVideo),
       Link,
       FontFamily,
       Mathematics,
@@ -146,7 +147,7 @@ const Tiptap = () => {
           }
         </code>
       </pre>
-      <img width="200" height="150" src="https://images.unsplash.com/photo-1682685796186-1bb4a5655653?ixlib=rb-4.0.3&ixid=M3wxMjA3fDF8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1740&q=80" alt="mon image" />
+      <img width="400" height="150" src="https://images.unsplash.com/photo-1682685796186-1bb4a5655653?ixlib=rb-4.0.3&ixid=M3wxMjA3fDF8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1740&q=80" alt="mon image" />
       <p>
         I know, I know, this is impressive. It's only the tip of the iceberg though. Give it a try and click a little bit around. Don't forget to check the other examples too.
       </p>
@@ -271,6 +272,7 @@ const Tiptap = () => {
                 src: `/workspace/document/${img._id}`,
                 alt: img.alt,
                 title: img.title,
+                ["media-type"]: "img",
               })
               .run();
           });
